Consolidate InstrumentSettings container/component logic, remove duplicate logic from ChooseOrCreateInstrumentDialog
